created objects of all the parts
